Fix: Dependency handling when converting on-run-start / on-run-end hooks in dbt projects
